ref(api): Move schema validation error handling out of `validate_request`
