+++ snuba/api.py
@@ -10,6 +10,8 @@ import sentry_sdk
 from sentry_sdk.integrations.flask import FlaskIntegration
 from sentry_sdk.integrations.gnu_backtrace import GnuBacktraceIntegration
 import simplejson as json
+from werkzeug.exceptions import BadRequest
+import jsonschema
 
 from snuba import schemas, settings, state, util
 from snuba.clickhouse import ClickhousePool
@@ -64,6 +66,32 @@ sentry_sdk.init(
     integrations=[FlaskIntegration(), GnuBacktraceIntegration()],
     release=os.getenv('SNUBA_RELEASE')
 )
+
+
+@application.errorhandler(BadRequest)
+def handle_bad_request(exception: BadRequest):
+    cause = getattr(exception, '__cause__')
+    if cause is not None and isinstance(cause, json.errors.JSONDecodeError):
+        data = {'error': {'type': 'json', 'message': str(cause)}}
+    elif cause is not None and isinstance(cause, jsonschema.ValidationError):
+        data = {'error': {
+            'type': 'schema',
+            'message': cause.message,
+            'path': list(cause.path),
+            'schema': cause.schema,
+        }}
+    else:
+        data = {'error': {'type': 'request', 'message': str(exception)}}
+
+    def default_encode(value):
+        # XXX: This is necessary for rendering schema defaults values that are
+        # generated by callables, rather than constants.
+        if callable(value):
+            return value()
+        else:
+            raise TypeError()
+
+    return json.dumps(data, sort_keys=True, indent=4, default=default_encode), 400, {'Content-Type': 'application/json'}
 
 
 @application.route('/')
